Write deletions first
A diff might mark an unique object for deletion in order to then
create a new object (for example a case change in a username in a LDAP
database with case insensitive and unique usernames). The deletion
operation then needs to happen first in order for the addition to not
fail with a conflict.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>